+++ CHANGELOG.md
@@ -7,10 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+### Changed
+- Updated `litellm` to `>=1.72.2`.
+- Updated `ollama` to `>=0.5.1`.
+
 ### Fixed
 - Fixed an issue where the EOS token would be included in the vLLM generation output,
   leading to incorrect evaluation results. We now manually remove all stop tokens from
   the generation output, which fixes this issue.
+- Now correctly detects reasoning models for Ollama models and enables their new "think"
+  parameter whenever a reasoning model is detected.
 
 
 ## [v15.9.2] - 2025-06-04